+++ src/denovo_molecular_formula_prediction.cpp
@@ -58,7 +58,7 @@ std::vector<std::string> DeNovoMolecularFormulaPrediction(const Rcpp::NumericVec
 	std::vector<std::string> resultantFormulas(size);
 	counter = 0;
 	Rcpp::Rcout << "Calculating fragmentation trees..." << std::endl;
-	for (size_t i = 0; i < size; i++) {
+	for (int i = 0; i < size; i++) {
 		FragmentationTree tree(allNodes[i], inputData[i].parentMass);
 		const int colorZeroFormulaCount = tree.GetColorZeroCount();
 		if (colorZeroFormulaCount <= 0) continue; // Mean there are no decompositions!
